moved api tests
