+++ test/api/PrintLog.js
@@ -1,24 +1,24 @@
 /* global describe it before after */
 // During the test the env variable is set to test
 process.env.NODE_ENV = 'test';
-const PrintLog = require('../api/peripheral_api/models/PrintLog');
+const PrintLog = require('../../api/peripheral_api/models/PrintLog');
 const chai = require('chai');
 const chaiHttp = require('chai-http');
-const constants = require('../api/util/constants');
+const constants = require('../../api/util/constants');
 const { OK, BAD_REQUEST } = constants.STATUS_CODES;
-const SceApiTester = require('../test/util/tools/SceApiTester');
+const SceApiTester = require('../../test/util/tools/SceApiTester');
 
 let app = null;
 let test = null;
 const expect = chai.expect;
-const tools = require('./util/tools/tools.js');
+const tools = require('../util/tools/tools.js');
 chai.should();
 chai.use(chaiHttp);
 
 describe('PrintLog', () => {
   before(done => {
     app = tools.initializeServer(
-      __dirname + '/../api/peripheral_api/routes/PrintLog.js');
+      __dirname + '/../../api/peripheral_api/routes/PrintLog.js');
     test = new SceApiTester(app);
     tools.emptySchema(PrintLog);
     done();
@@ -49,13 +49,13 @@ describe('PrintLog', () => {
     });
 
     it('Should return statusCode 200 when all required ' +
-    'fields are filled in', async () => {
-      const result = await test.sendPostRequest(
-        '/api/PrintLog/addPrintLog',
-        VALID_PRINT_LOG
-      );
-      expect(result).to.have.status(OK);
-    });
+      'fields are filled in', async () => {
+        const result = await test.sendPostRequest(
+          '/api/PrintLog/addPrintLog',
+          VALID_PRINT_LOG
+        );
+        expect(result).to.have.status(OK);
+      });
   });
 
   describe('/GET getPrintLogs', () => {
