+++ test/api/S3Bucket.js
@@ -3,18 +3,18 @@ process.env.NODE_ENV = 'test';
 const chai = require('chai');
 const chaiHttp = require('chai-http');
 const sinon = require('sinon');
-const constants = require('../api/util/constants');
+const constants = require('../../api/util/constants');
 const { OK, NOT_FOUND } = constants.STATUS_CODES;
-const SceApiTester = require('./util/tools/SceApiTester');
+const SceApiTester = require('../util/tools/SceApiTester');
 const {
   S3BucketApiHandler,
-} = require('../api/cloud_api/util/S3BucketApiHandler');
+} = require('../../api/cloud_api/util/S3BucketApiHandler');
 
 let app = null;
 let test = null;
 let sandbox = sinon.createSandbox();
 const expect = chai.expect;
-const tools = require('./util/tools/tools.js');
+const tools = require('../util/tools/tools.js');
 
 chai.should();
 chai.use(chaiHttp);
@@ -24,7 +24,7 @@ describe('S3Bucket', () => {
   before((done) => {
     bucketStub = sandbox.stub(S3BucketApiHandler.prototype, 'getSignedUrl');
     app = tools.initializeServer(
-      __dirname + '/../api/cloud_api/routes/S3Bucket.js'
+      __dirname + '/../../api/cloud_api/routes/S3Bucket.js'
     );
     test = new SceApiTester(app);
     done();
