+++ test/api/Sheets.js
@@ -3,17 +3,17 @@ process.env.NODE_ENV = 'test';
 const chai = require('chai');
 const chaiHttp = require('chai-http');
 const sinon = require('sinon');
-const constants = require('../api/util/constants');
+const constants = require('../../api/util/constants');
 const { OK, NOT_FOUND } = constants.STATUS_CODES;
-const SceApiTester = require('./util/tools/SceApiTester');
+const SceApiTester = require('../util/tools/SceApiTester');
 const { SceGoogleApiHandler } =
-  require('../api/cloud_api/util/SceGoogleApiHandler');
+  require('../../api/cloud_api/util/SceGoogleApiHandler');
 
 let app = null;
 let test = null;
 let sandbox = sinon.createSandbox();
 const expect = chai.expect;
-const tools = require('./util/tools/tools.js');
+const tools = require('../util/tools/tools.js');
 
 
 chai.should();
@@ -24,7 +24,7 @@ describe('Sheets', () => {
   before(done => {
     sheetsStub = sandbox.stub(SceGoogleApiHandler.prototype, 'writeToForm');
     app = tools.initializeServer(
-      __dirname + '/../api/cloud_api/routes/Sheets.js');
+      __dirname + '/../../api/cloud_api/routes/Sheets.js');
     test = new SceApiTester(app);
     done();
   });
@@ -49,7 +49,7 @@ describe('Sheets', () => {
 
   const SHEETS_RESPONSE = 'Sample Response';
 
-  describe('/POST addToSpreadsheet', () =>{
+  describe('/POST addToSpreadsheet', () => {
     it('Should return 200 when a row of data is successfully added',
       async () => {
         sheetsStub.resolves(SHEETS_RESPONSE);
