+++ test/api/User.js
@@ -2,7 +2,7 @@
 // During the test the env variable is set to test
 process.env.NODE_ENV = 'test';
 
-const User = require('../api/main_endpoints/models/User');
+const User = require('../../../api/main_endpoints/models/User.js');
 
 // Require the dev-dependencies
 const chai = require('chai');
@@ -13,10 +13,10 @@ const {
   UNAUTHORIZED,
   NOT_FOUND,
   FORBIDDEN
-} = require('../api/util/constants').STATUS_CODES;
+} = require('../../../api/util/constants').STATUS_CODES;
 const sinon = require('sinon');
-const SceApiTester = require('./util/tools/SceApiTester');
-const discordModule = require('../api/main_endpoints/util/discord-connection');
+const SceApiTester = require('../../util/tools/SceApiTester');
+const discordModule = require('../../../api/main_endpoints/util/discord-connection');
 
 
 let app = null;
@@ -24,13 +24,13 @@ let test = null;
 let sandbox = sinon.createSandbox();
 
 const expect = chai.expect;
-const tools = require('./util/tools/tools.js');
+const tools = require('../../util/tools/tools.js');
 const {
   setTokenStatus,
   resetTokenMock,
   restoreTokenMock,
   initializeTokenMock
-} = require('./util/mocks/TokenValidFunctions');
+} = require('../../util/mocks/TokenValidFunctions');
 
 chai.should();
 chai.use(chaiHttp);
@@ -40,8 +40,8 @@ describe('User', () => {
   before(done => {
     initializeTokenMock();
     app = tools.initializeServer([
-      __dirname + '/../api/main_endpoints/routes/User.js',
-      __dirname + '/../api/main_endpoints/routes/Auth.js'
+      __dirname + '/../../api/main_endpoints/routes/User.js',
+      __dirname + '/../../api/main_endpoints/routes/Auth.js'
     ]);
     test = new SceApiTester(app);
     // Before each test we empty the database
@@ -66,12 +66,12 @@ describe('User', () => {
 
   describe('/POST checkIfUserExists with no users added yet', () => {
     it('Should return statusCode 400 when an email is not' +
-     'provided', async () => {
-      const user = {};
-      const result = await test.sendPostRequest(
-        '/api/User/checkIfUserExists', user);
-      expect(result).to.have.status(BAD_REQUEST);
-    });
+      'provided', async () => {
+        const user = {};
+        const result = await test.sendPostRequest(
+          '/api/User/checkIfUserExists', user);
+        expect(result).to.have.status(BAD_REQUEST);
+      });
 
     it('Should return statusCode 200 when a user does not exist', async () => {
       const user = {
@@ -101,25 +101,25 @@ describe('User', () => {
     });
 
     it('Should return statusCode 401 if an invalid ' +
-       'token was passed in', async () => {
-      const user = {
-        token: 'Invalid token'
-      };
-      const result = await test.sendPostRequest(
-        '/api/User/users', user);
-      expect(result).to.have.status(UNAUTHORIZED);
-    });
+      'token was passed in', async () => {
+        const user = {
+          token: 'Invalid token'
+        };
+        const result = await test.sendPostRequest(
+          '/api/User/users', user);
+        expect(result).to.have.status(UNAUTHORIZED);
+      });
 
     it('Should return statusCode 200 and return an array ' +
-       'of all objects in collection', async () => {
-      const form = {
-        token: token
-      };
-      setTokenStatus(true);
-      const result = await test.sendPostRequestWithToken(
-        token, '/api/User/users', form);
-      expect(result).to.have.status(OK);
-    });
+      'of all objects in collection', async () => {
+        const form = {
+          token: token
+        };
+        setTokenStatus(true);
+        const result = await test.sendPostRequestWithToken(
+          token, '/api/User/users', form);
+        expect(result).to.have.status(OK);
+      });
   });
 
   describe('/POST searchFor', () => {
@@ -133,15 +133,15 @@ describe('User', () => {
     });
 
     it('Should return statusCode 401 if an invalid ' +
-       'token was passed in', async () => {
-      const user = {
-        email: 'a@b.c',
-        token: 'Invalid token'
-      };
-      const result = await test.sendPostRequest(
-        '/api/User/search', user);
-      expect(result).to.have.status(UNAUTHORIZED);
-    });
+      'token was passed in', async () => {
+        const user = {
+          email: 'a@b.c',
+          token: 'Invalid token'
+        };
+        const result = await test.sendPostRequest(
+          '/api/User/search', user);
+        expect(result).to.have.status(UNAUTHORIZED);
+      });
 
     it('Should return statusCode 404 if no user was found', async () => {
       const user = {
@@ -155,27 +155,27 @@ describe('User', () => {
     });
 
     it('Should return statusCode 200 and a user if ' +
-       'the query was found', async () => {
-      const user = {
-        email: 'a@b.c',
-        token: token
-      };
-      setTokenStatus(true);
-      const result = await test.sendPostRequestWithToken(
-        token, '/api/User/search', user);
-      expect(result).to.have.status(OK);
-      result.body.should.be.a('object');
-      result.body.should.have.property('firstName');
-      result.body.should.have.property('lastName');
-      result.body.should.have.property('email');
-      result.body.should.have.property('emailVerified');
-      result.body.should.have.property('emailOptIn');
-      result.body.should.have.property('accessLevel');
-      result.body.should.have.property('major');
-      result.body.should.have.property('joinDate');
-      result.body.should.have.property('lastLogin');
-      result.body.should.have.property('discordID');
-    });
+      'the query was found', async () => {
+        const user = {
+          email: 'a@b.c',
+          token: token
+        };
+        setTokenStatus(true);
+        const result = await test.sendPostRequestWithToken(
+          token, '/api/User/search', user);
+        expect(result).to.have.status(OK);
+        result.body.should.be.a('object');
+        result.body.should.have.property('firstName');
+        result.body.should.have.property('lastName');
+        result.body.should.have.property('email');
+        result.body.should.have.property('emailVerified');
+        result.body.should.have.property('emailOptIn');
+        result.body.should.have.property('accessLevel');
+        result.body.should.have.property('major');
+        result.body.should.have.property('joinDate');
+        result.body.should.have.property('lastLogin');
+        result.body.should.have.property('discordID');
+      });
   });
 
   describe('/POST edit', () => {
@@ -189,15 +189,15 @@ describe('User', () => {
     });
 
     it('Should return statusCode 401 if an invalid ' +
-       'token was passed in', async () => {
-      const user = {
-        email: 'a@b.c',
-        token: 'Invalid token'
-      };
-      const result = await test.sendPostRequest(
-        '/api/User/edit', user);
-      expect(result).to.have.status(UNAUTHORIZED);
-    });
+      'token was passed in', async () => {
+        const user = {
+          email: 'a@b.c',
+          token: 'Invalid token'
+        };
+        const result = await test.sendPostRequest(
+          '/api/User/edit', user);
+        expect(result).to.have.status(UNAUTHORIZED);
+      });
 
     it('Should return statusCode 404 if no user was found', async () => {
       const user = {
@@ -211,20 +211,20 @@ describe('User', () => {
     });
 
     it('Should return statusCode 200 and a message ' +
-       'if a user was edited', async () => {
-      const user = {
-        email: 'a@b.c',
-        token: token,
-        firstName: 'pinkUnicorn',
-        numberOfSemestersToSignUpFor: undefined
-      };
-      setTokenStatus(true);
-      const result = await test.sendPostRequestWithToken(
-        token, '/api/User/edit', user);
-      expect(result).to.have.status(OK);
-      result.body.should.be.a('object');
-      result.body.should.have.property('message');
-    });
+      'if a user was edited', async () => {
+        const user = {
+          email: 'a@b.c',
+          token: token,
+          firstName: 'pinkUnicorn',
+          numberOfSemestersToSignUpFor: undefined
+        };
+        setTokenStatus(true);
+        const result = await test.sendPostRequestWithToken(
+          token, '/api/User/edit', user);
+        expect(result).to.have.status(OK);
+        result.body.should.be.a('object');
+        result.body.should.have.property('message');
+      });
   });
 
   describe('/POST delete', () => {
@@ -238,15 +238,15 @@ describe('User', () => {
     });
 
     it('Should return statusCode 403 if an invalid ' +
-       'token was passed in', async () => {
-      const user = {
-        email: 'a@b.c',
-        token: 'Invalid token'
-      };
-      const result = await test.sendPostRequest(
-        '/api/User/delete', user);
-      expect(result).to.have.status(UNAUTHORIZED);
-    });
+      'token was passed in', async () => {
+        const user = {
+          email: 'a@b.c',
+          token: 'Invalid token'
+        };
+        const result = await test.sendPostRequest(
+          '/api/User/delete', user);
+        expect(result).to.have.status(UNAUTHORIZED);
+      });
 
     it('Should return statusCode 404 if no user was found', async () => {
       const user = {
@@ -260,18 +260,18 @@ describe('User', () => {
     });
 
     it('Should return statusCode 200 and a message ' +
-       'if a user was deleted', async () => {
-      const user = {
-        email: 'a@b.c',
-        token: token
-      };
-      setTokenStatus(true);
-      const result = await test.sendPostRequestWithToken(
-        token, '/api/User/delete', user);
-      expect(result).to.have.status(OK);
-      result.body.should.be.a('object');
-      result.body.should.have.property('message');
-    });
+      'if a user was deleted', async () => {
+        const user = {
+          email: 'a@b.c',
+          token: token
+        };
+        setTokenStatus(true);
+        const result = await test.sendPostRequestWithToken(
+          token, '/api/User/delete', user);
+        expect(result).to.have.status(OK);
+        result.body.should.be.a('object');
+        result.body.should.have.property('message');
+      });
   });
 
   describe('/POST connectToDiscord', () => {
@@ -285,35 +285,35 @@ describe('User', () => {
     });
     it('Should return statusCode 401 if an invalid ' +
       'token was passed in', async () => {
-      const user = {
-        email: 'a@b.c',
-        token: 'Invalid token'
-      };
-      const result = await test.sendPostRequest(
-        '/api/user/connectToDiscord', user);
-      expect(result).to.have.status(UNAUTHORIZED);
-    });
+        const user = {
+          email: 'a@b.c',
+          token: 'Invalid token'
+        };
+        const result = await test.sendPostRequest(
+          '/api/user/connectToDiscord', user);
+        expect(result).to.have.status(UNAUTHORIZED);
+      });
     it('Should return statusCode 400 if an incorrect or no ' +
       'email was used', async () => {
-      const user = {
-        token
-      };
-      setTokenStatus(true);
-      const result = await test.sendPostRequestWithToken(
-        token, '/api/user/connectToDiscord', user);
-      expect(result).to.have.status(BAD_REQUEST);
-    });
+        const user = {
+          token
+        };
+        setTokenStatus(true);
+        const result = await test.sendPostRequestWithToken(
+          token, '/api/user/connectToDiscord', user);
+        expect(result).to.have.status(BAD_REQUEST);
+      });
     it('Should return statusCode 200 ' +
       'if Discord connection was successful', async () => {
-      const user = {
-        email: 'a@b.c',
-        token
-      };
-      setTokenStatus(true);
-      const result = await test.sendPostRequestWithToken(
-        token, '/api/user/connectToDiscord', user);
-      expect(result).to.have.status(OK);
-    });
+        const user = {
+          email: 'a@b.c',
+          token
+        };
+        setTokenStatus(true);
+        const result = await test.sendPostRequestWithToken(
+          token, '/api/user/connectToDiscord', user);
+        expect(result).to.have.status(OK);
+      });
   });
 
   describe('/GET callback', () => {
