+++ test/api/officerManager.js
@@ -1,8 +1,8 @@
 /* global describe it before after beforeEach afterEach */
 // During the test the env variable is set to test
 process.env.NODE_ENV = 'test';
-const OfficerManager = require('../api/main_endpoints/models/OfficerManager');
-const User = require('../api/main_endpoints/models/User');
+const OfficerManager = require('../../api/main_endpoints/models/OfficerManager');
+const User = require('../../api/main_endpoints/models/User');
 const chai = require('chai');
 const chaiHttp = require('chai-http');
 const {
@@ -11,21 +11,21 @@ const {
   UNAUTHORIZED,
   NOT_FOUND,
   FORBIDDEN
-} = require('../api/util/constants').STATUS_CODES;
-const SceApiTester = require('../test/util/tools/SceApiTester');
+} = require('../../api/util/constants').STATUS_CODES;
+const SceApiTester = require('../../test/util/tools/SceApiTester');
 
 let app = null;
 let test = null;
 const expect = chai.expect;
 // tools for testing
-const tools = require('./util/tools/tools.js');
+const tools = require('../util/tools/tools.js');
 const {
   setTokenStatus,
   resetTokenMock,
   restoreTokenMock,
   initializeTokenMock
-} = require('./util/mocks/TokenValidFunctions');
-const { DEFAULT_PHOTO_URL } = require('../api/util/constants');
+} = require('../util/mocks/TokenValidFunctions');
+const { DEFAULT_PHOTO_URL } = require('../../api/util/constants');
 chai.should();
 chai.use(chaiHttp);
 
@@ -34,7 +34,7 @@ describe('OfficerManager', () => {
   before(done => {
     initializeTokenMock();
     app = tools.initializeServer(
-      __dirname + '/../api/main_endpoints/routes/officerManager.js');
+      __dirname + '/../../api/main_endpoints/routes/officerManager.js');
     test = new SceApiTester(app);
     // Before each test we empty the database
     tools.emptySchema(OfficerManager);
@@ -56,13 +56,13 @@ describe('OfficerManager', () => {
 
   describe('/POST submit', () => {
     it('Should return statusCode 400 when the ' +
-       'required fields are not set', async () => {
-      const form = { token: token };
-      setTokenStatus(true);
-      const result = await test.sendPostRequest(
-        '/api/officerManager/submit', form);
-      expect(result).to.have.status(BAD_REQUEST);
-    });
+      'required fields are not set', async () => {
+        const form = { token: token };
+        setTokenStatus(true);
+        const result = await test.sendPostRequest(
+          '/api/officerManager/submit', form);
+        expect(result).to.have.status(BAD_REQUEST);
+      });
 
     it('Should return statusCode 403 when no token is submited', async () => {
       const form = {
@@ -87,22 +87,22 @@ describe('OfficerManager', () => {
       });
 
     it('Should return statusCode 200 when all required ' +
-       'fields are filled in', async () => {
-      const form = {
-        email: 'test@test.com',
-        team: 'dev',
-        facebook: 'facebooklink',
-        github: 'githublink',
-        linkedin: 'linkedinlink',
-        quote: 'aquote',
-        pictureUrl: DEFAULT_PHOTO_URL,
-        token: token
-      };
-      setTokenStatus(true);
-      const result = await test.sendPostRequest(
-        '/api/officerManager/submit', form);
-      expect(result).to.have.status(OK);
-    });
+      'fields are filled in', async () => {
+        const form = {
+          email: 'test@test.com',
+          team: 'dev',
+          facebook: 'facebooklink',
+          github: 'githublink',
+          linkedin: 'linkedinlink',
+          quote: 'aquote',
+          pictureUrl: DEFAULT_PHOTO_URL,
+          token: token
+        };
+        setTokenStatus(true);
+        const result = await test.sendPostRequest(
+          '/api/officerManager/submit', form);
+        expect(result).to.have.status(OK);
+      });
   });
 
   describe('/POST GetForm', () => {
@@ -171,15 +171,15 @@ describe('OfficerManager', () => {
     });
 
     it('Should return statusCode 401 if an invalid ' +
-       'token was passed in', async () => {
-      const form = {
-        token: 'Invalid token',
-        email: 'test@test.com'
-      };
-      const result = await test.sendPostRequest(
-        '/api/officerManager/edit', form);
-      expect(result).to.have.status(UNAUTHORIZED);
-    });
+      'token was passed in', async () => {
+        const form = {
+          token: 'Invalid token',
+          email: 'test@test.com'
+        };
+        const result = await test.sendPostRequest(
+          '/api/officerManager/edit', form);
+        expect(result).to.have.status(UNAUTHORIZED);
+      });
 
     it('Should return statusCode 404 if no form was found', async () => {
       const form = {
@@ -193,19 +193,19 @@ describe('OfficerManager', () => {
     });
 
     it('Should return statusCode 200 and a message ' +
-       'if a form was edited', async () => {
-      const form = {
-        token: token,
-        email: 'test@test.com',
-        team: 'event'
-      };
-      setTokenStatus(true);
-      const result = await test.sendPostRequest(
-        '/api/officerManager/edit', form);
-      expect(result).to.have.status(OK);
-      result.body.should.be.a('object');
-      result.body.should.have.property('message');
-    });
+      'if a form was edited', async () => {
+        const form = {
+          token: token,
+          email: 'test@test.com',
+          team: 'event'
+        };
+        setTokenStatus(true);
+        const result = await test.sendPostRequest(
+          '/api/officerManager/edit', form);
+        expect(result).to.have.status(OK);
+        result.body.should.be.a('object');
+        result.body.should.have.property('message');
+      });
 
     it('The update should be reflected in the database', async () => {
       const form = {
@@ -232,15 +232,15 @@ describe('OfficerManager', () => {
     });
 
     it('Should return statusCode 401 if an invalid ' +
-       'token was passed in', async () => {
-      const form = {
-        token: 'Invalid token',
-        email: 'test@test.com'
-      };
-      const result = await test.sendPostRequest(
-        '/api/officerManager/delete', form);
-      expect(result).to.have.status(UNAUTHORIZED);
-    });
+      'token was passed in', async () => {
+        const form = {
+          token: 'Invalid token',
+          email: 'test@test.com'
+        };
+        const result = await test.sendPostRequest(
+          '/api/officerManager/delete', form);
+        expect(result).to.have.status(UNAUTHORIZED);
+      });
 
     it('Should return statusCode 404 if no form was found', async () => {
       const form = {
@@ -254,18 +254,18 @@ describe('OfficerManager', () => {
     });
 
     it('Should return statusCode 200 and a message ' +
-       'if a form was deleted', async () => {
-      const form = {
-        token: token,
-        email: 'test@test.com'
-      };
-      setTokenStatus(true);
-      const result = await test.sendPostRequest(
-        '/api/officerManager/delete', form);
-      expect(result).to.have.status(OK);
-      result.body.should.be.a('object');
-      result.body.should.have.property('message');
-    });
+      'if a form was deleted', async () => {
+        const form = {
+          token: token,
+          email: 'test@test.com'
+        };
+        setTokenStatus(true);
+        const result = await test.sendPostRequest(
+          '/api/officerManager/delete', form);
+        expect(result).to.have.status(OK);
+        result.body.should.be.a('object');
+        result.body.should.have.property('message');
+      });
 
     it('The deleted item should be reflected in the database', async () => {
       const form = {
